# opt: skip number in json.Number  parsing