--- a/internal/native/avx/native_subr_amd64.go
+++ b/internal/native/avx/native_subr_amd64.go
@@ -9,45 +9,45 @@ package avx
 func __native_entry__() uintptr
 
 var (
-    _subr__f64toa       = __native_entry__() + 570
-    _subr__html_escape  = __native_entry__() + 8304
-    _subr__i64toa       = __native_entry__() + 3653
-    _subr__lspace       = __native_entry__() + 251
+    _subr__f64toa       = __native_entry__() + 630
+    _subr__html_escape  = __native_entry__() + 8160
+    _subr__i64toa       = __native_entry__() + 3642
+    _subr__lspace       = __native_entry__() + 301
     _subr__lzero        = __native_entry__() + 13
-    _subr__quote        = __native_entry__() + 4970
-    _subr__skip_array   = __native_entry__() + 17520
-    _subr__skip_number  = __native_entry__() + 20459
-    _subr__skip_object  = __native_entry__() + 17557
-    _subr__skip_one     = __native_entry__() + 15672
-    _subr__u64toa       = __native_entry__() + 3748
-    _subr__unquote      = __native_entry__() + 6010
-    _subr__validate_one = __native_entry__() + 20576
-    _subr__value        = __native_entry__() + 10907
-    _subr__vnumber      = __native_entry__() + 13686
-    _subr__vsigned      = __native_entry__() + 15116
-    _subr__vstring      = __native_entry__() + 12639
-    _subr__vunsigned    = __native_entry__() + 15396
+    _subr__quote        = __native_entry__() + 4955
+    _subr__skip_array   = __native_entry__() + 17551
+    _subr__skip_number  = __native_entry__() + 20669
+    _subr__skip_object  = __native_entry__() + 17588
+    _subr__skip_one     = __native_entry__() + 15699
+    _subr__u64toa       = __native_entry__() + 3735
+    _subr__unquote      = __native_entry__() + 6005
+    _subr__validate_one = __native_entry__() + 20786
+    _subr__value        = __native_entry__() + 10880
+    _subr__vnumber      = __native_entry__() + 13857
+    _subr__vsigned      = __native_entry__() + 15171
+    _subr__vstring      = __native_entry__() + 12822
+    _subr__vunsigned    = __native_entry__() + 15430
 )
 
 const (
     _stack__f64toa = 120
-    _stack__html_escape = 64
+    _stack__html_escape = 72
     _stack__i64toa = 24
     _stack__lspace = 8
     _stack__lzero = 8
     _stack__quote = 80
-    _stack__skip_array = 152
-    _stack__skip_number = 88
-    _stack__skip_object = 152
-    _stack__skip_one = 152
+    _stack__skip_array = 160
+    _stack__skip_number = 96
+    _stack__skip_object = 160
+    _stack__skip_one = 160
     _stack__u64toa = 8
-    _stack__unquote = 80
-    _stack__validate_one = 152
-    _stack__value = 408
-    _stack__vnumber = 320
+    _stack__unquote = 88
+    _stack__validate_one = 160
+    _stack__value = 416
+    _stack__vnumber = 312
     _stack__vsigned = 16
-    _stack__vstring = 120
-    _stack__vunsigned = 16
+    _stack__vstring = 128
+    _stack__vunsigned = 8
 )
 
 var (