--- a/internal/native/avx2/native_amd64.go
+++ b/internal/native/avx2/native_amd64.go
@@ -107,4 +107,9 @@ func __skip_object(s *string, p *int, m *types.StateMachine) (ret int)
 //go:nosplit
 //go:noescape
 //goland:noinspection GoUnusedParameter
+func __skip_number(s *string, p *int) (ret int)
+
+//go:nosplit
+//go:noescape
+//goland:noinspection GoUnusedParameter
 func __validate_one(s *string, p *int, m *types.StateMachine) (ret int)