--- a/internal/native/avx2/native_export_amd64.go
+++ b/internal/native/avx2/native_export_amd64.go
@@ -42,4 +42,5 @@ var (
     S_skip_one    = _subr__skip_one
     S_skip_array  = _subr__skip_array
     S_skip_object = _subr__skip_object
+    S_skip_number = _subr__skip_number
 )