--- a/internal/native/avx2/native_subr_amd64.go
+++ b/internal/native/avx2/native_subr_amd64.go
@@ -9,24 +9,24 @@ package avx2
 func __native_entry__() uintptr
 
 var (
-    _subr__f64toa       = __native_entry__() + 825
-    _subr__html_escape  = __native_entry__() + 9653
-    _subr__i64toa       = __native_entry__() + 3908
-    _subr__lspace       = __native_entry__() + 379
+    _subr__f64toa       = __native_entry__() + 903
+    _subr__html_escape  = __native_entry__() + 9535
+    _subr__i64toa       = __native_entry__() + 3915
+    _subr__lspace       = __native_entry__() + 429
     _subr__lzero        = __native_entry__() + 13
-    _subr__quote        = __native_entry__() + 5325
-    _subr__skip_array   = __native_entry__() + 20295
-    _subr__skip_number  = __native_entry__() + 23815
-    _subr__skip_object  = __native_entry__() + 20332
-    _subr__skip_one     = __native_entry__() + 18452
-    _subr__u64toa       = __native_entry__() + 4003
-    _subr__unquote      = __native_entry__() + 7158
-    _subr__validate_one = __native_entry__() + 23932
-    _subr__value        = __native_entry__() + 13550
-    _subr__vnumber      = __native_entry__() + 16466
-    _subr__vsigned      = __native_entry__() + 17896
-    _subr__vstring      = __native_entry__() + 15559
-    _subr__vunsigned    = __native_entry__() + 18176
+    _subr__quote        = __native_entry__() + 5328
+    _subr__skip_array   = __native_entry__() + 21558
+    _subr__skip_number  = __native_entry__() + 25206
+    _subr__skip_object  = __native_entry__() + 21595
+    _subr__skip_one     = __native_entry__() + 18458
+    _subr__u64toa       = __native_entry__() + 4008
+    _subr__unquote      = __native_entry__() + 7080
+    _subr__validate_one = __native_entry__() + 25323
+    _subr__value        = __native_entry__() + 13781
+    _subr__vnumber      = __native_entry__() + 16616
+    _subr__vsigned      = __native_entry__() + 17930
+    _subr__vstring      = __native_entry__() + 15739
+    _subr__vunsigned    = __native_entry__() + 18189
 )
 
 const (
@@ -35,19 +35,19 @@ const (
     _stack__i64toa = 24
     _stack__lspace = 8
     _stack__lzero = 8
-    _stack__quote = 80
-    _stack__skip_array = 160
+    _stack__quote = 64
+    _stack__skip_array = 136
     _stack__skip_number = 96
-    _stack__skip_object = 160
-    _stack__skip_one = 160
+    _stack__skip_object = 136
+    _stack__skip_one = 136
     _stack__u64toa = 8
     _stack__unquote = 72
-    _stack__validate_one = 160
+    _stack__validate_one = 136
     _stack__value = 408
-    _stack__vnumber = 320
+    _stack__vnumber = 312
     _stack__vsigned = 16
     _stack__vstring = 112
-    _stack__vunsigned = 16
+    _stack__vunsigned = 8
 )
 
 var (