--- a/internal/native/native_export_amd64.tmpl
+++ b/internal/native/native_export_amd64.tmpl
@@ -40,4 +40,5 @@ var (
     S_skip_one    = _subr__skip_one
     S_skip_array  = _subr__skip_array
     S_skip_object = _subr__skip_object
+    S_skip_number = _subr__skip_number
 )