--- a/native/native.h
+++ b/native/native.h
@@ -108,7 +108,7 @@ ssize_t quote(const char *sp, ssize_t nb, char *dp, ssize_t *dn, uint64_t flags)
 ssize_t unquote(const char *sp, ssize_t nb, char *dp, ssize_t *ep, uint64_t flags);
 ssize_t html_escape(const char *sp, ssize_t nb, char *dp, ssize_t *dn);
 
-long value(const char *s, size_t n, long p, JsonState *ret, int allow_control);
+long value(const char *s, size_t n, long p, JsonState *ret, uint64_t flags);
 void vstring(const GoString *src, long *p, JsonState *ret);
 void vnumber(const GoString *src, long *p, JsonState *ret);
 void vsigned(const GoString *src, long *p, JsonState *ret);