… internal/native/avx/native_amd64.s
… internal/native/avx2/native_amd64.s