use skip_number instead of parsing when decoding `json.Number`, because `json.Number` is actually a string of digits.