@AsterDY 2022-04-01
### Benchmarks
- Before
```
go116 test -benchmem -run=^$ -benchtime=10000000x -bench "BenchmarkDecodeJsonNumber_Sonic|BenchmarkDecodeUseNumber_Sonic"
goos: darwin
goarch: amd64
pkg: github.com/bytedance/sonic/issue_test
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkDecodeJsonNumber_Sonic-16      10000000              2261 ns/op          49.97 MB/s          18 B/op          1 allocs/op
BenchmarkDecodeUseNumber_Sonic-16       10000000               120.0 ns/op       942.03 MB/s          41 B/op          2 allocs/op
```
- After:
```
go116 test -benchmem -run=^$ -benchtime=10000000x -bench "BenchmarkDecodeJsonNumber_Sonic|BenchmarkDecodeUseNumber_Sonic"
goos: darwin
goarch: amd64
pkg: github.com/bytedance/sonic/issue_test
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkDecodeJsonNumber_Sonic-16      10000000                78.14 ns/op     1446.14 MB/s          23 B/op          1 allocs/op
BenchmarkDecodeUseNumber_Sonic-16       10000000               110.4 ns/op      1023.17 MB/s          41 B/op          2 allocs/op
```